chore: move barButtonItem to its own file
